pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "debugpy>=1.8.14,<2",
  "isort>=6.0.1,<7",
  "ab-test-fixtures>=0.1.5,<0.2.0",
  "pre-commit>=4.2.0,<5",
  "pytest-asyncio>=1.0.0,<2",
  "pytest-cov>=6.2.1,<7",
  "pytest>=8.4.1,<9",
  "ruff>=0.12.3,<0.13",
  "tox>=4.27.0,<5"
]

[project]
authors = [{email = "mattcoul7@gmail.com", name = "Matt Coulter"}]
dependencies = []
description = "A template package template."
name = "ab-template-package"
readme = "README.md"
requires-python = ">=3.12,<4.0"
version = "0.1.0"

[tool.hatch.build.targets.sdist]
include = ["src/ab_core"]

[tool.hatch.build.targets.wheel]
include = ["src/ab_core"]

[tool.hatch.build.targets.wheel.sources]
"src/ab_core" = "ab_core"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
markers = [
  "integration: Assigns the test to the integration test suite.",
  "regression: Assigns the test to the regression test suite"
]

[tool.ruff]
line-length = 120
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
exclude = [".git", ".venv", "__pycache__", "proto"]
fixable = ["ALL"]
ignore = [
  "D104" # missing docstring in public package
]
select = [
  "ARG001", # unused arguments
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "D", # pydocstyle (docstring checks)
  "E", # pycodestyle errors
  "F", # pyflakes
  "I", # isort
  "UP", # pyupgrade
  "W" # pycodestyle warnings
]
